♻️ split the changelog script in smaller files
