+++ scripts/release/generate-changelog/lib/constants.js
@@ -1,0 +1,17 @@
+exports.CONTRIBUTING_FILE = 'CONTRIBUTING.md'
+exports.CHANGELOG_FILE = 'CHANGELOG.md'
+exports.PUBLIC_EMOJI_PRIORITY = ['💥', '✨', '🐛', '⚡', '📝']
+exports.INTERNAL_EMOJI_PRIORITY = [
+  '👷',
+  '🔧',
+  '📦', // build conf
+  '♻️',
+  '🎨', // refactoring
+  '🧪',
+  '✅', // tests
+  '🔇',
+  '🔊', // telemetry
+  '👌',
+  '📄',
+  '⚗️', // experiment
+]
